Average speed decoration over larger amount of timestamps
This needs to be further modified, different ways of calculating the
distance, speed and averaging need to be implemented.

Ideally it would take the resolution of the terminal as a good
indicator to what integration would be necessary. However there is
more work involved than just taking the direct distance from point

 x_n to x_{n+d}

as the route might not be direct between the two.